Suppression de l'ancien contenu de la fenêtre de récompenses
