+++ jeux/Jeu_1/Jeu.py
@@ -14,7 +14,7 @@ from jeux.Jeu_1.action.Placement import Placement
 from jeux.Jeu_1.ui.fleche import Fleche
 from jeux.Jeu_1.ui.barreAction import BarreAction
 from jeux.Jeu_1.orgaFen import OrgaFen
-from jeux.Jeu_1.recompFen import RecompFen
+from jeux.Jeu_1.recompense.recompFen import RecompFen
 from jeux.Jeu_1.fenFin.fin import Fin
 from museeNoyee import orga, miniorga, abordage, miniabordage
